Remove archived code
