+++ .gitignore
@@ -121,3 +121,4 @@ app.log
 callback_audit_results/
 
 !load-tests/thresholds.json
+archive/
