--- archive/legacy_backend/backend/requirements.txt
@@ -1,5 +1,0 @@
-flask==2.3.2
-flask-cors==4.0.0
-pandas==2.0.3
-openpyxl==3.1.2
-werkzeug==2.3.6
