--- archive/mvp/__init__.py
@@ -1,1 +1,0 @@
-"""Minimal MVP package for independent pipeline."""
